+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Reissue.finalize returns the version and value as an array
 - Documentation on the refined redo method in Gem::Version
 - Limit major numbers to Integers
+- Handle empty changelog files
 
 ## [0.1.2] - 2024-06-08